requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
aiosmtpd==1.4.6
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-timeout==5.0.0
atpublic==5.0
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
bcrypt==4.0.1
boto3==1.35.50
botocore==1.35.50
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
decorator==4.4.2
distlib==0.3.9
dnspython==2.6.1
email_validator==2.2.0
etelemetry==0.3.1
exceptiongroup==1.2.2
executing==2.1.0
factory_boy==3.3.1
Faker==28.4.1
fastapi==0.114.2
fastapi-admin==1.0.4
fastapi-cli==0.0.5
fastapi-pagination==0.12.31
fastapi-utils==0.7.0
filelock==3.16.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
identify==2.6.1
idna==3.9
imageio==2.36.0
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
ipython==8.6.0
iso8601==2.1.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
looseversion==1.3.0
lxml==5.3.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
moviepy==1.0.3
mypy-extensions==1.0.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.8.6
nodeenv==1.9.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pathlib==1.0.1
pdf2image==1.17.0
pendulum==3.0.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.5.0
proglog==0.1.10
prompt-toolkit==3.0.30
prov==2.0.1
psutil==5.9.8
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
pydot==3.0.2
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.13
pyparsing==3.2.0
pypika-tortoise==0.2.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
requests==2.32.3
responses==0.25.3
rich==13.8.1
s3transfer==0.10.3
scipy==1.14.1
sentry-sdk==2.17.0
setuptools==75.3.0
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sqladmin==0.19.0
SQLAlchemy==2.0.34
sqlalchemy_database==0.1.2
sqlmodel==0.0.22
stack-data==0.6.3
starlette==0.38.5
time-machine==2.16.0
tomli==2.0.1
tortoise-orm==0.21.7
tqdm==4.66.6
traitlets==5.14.3
traits==6.3.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.26.6
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
WTForms==3.1.2
